Consolidate UseTrusteeFormValidation types into UseTrusteeForm
Jira ticket: CAMS-584
